% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_am_par.R
\name{abc_am_par}
\alias{abc_am_par}
\title{Adaptive ABC-MCMC}
\usage{
abc_am_par(
  obs,
  tol,
  kernel_func,
  p_s_parallel,
  prior,
  theta_0,
  s_0,
  n_adapt,
  n_iter,
  sigma_s,
  sigma_0,
  sigma_epi,
  string
)
}
\arguments{
\item{obs}{The vector of observed data point.}

\item{tol}{A positive numeric value for the tolerance.}

\item{kernel_func}{A kernel function.}

\item{p_s_parallel}{A function to provide the sampled summary statistics by given parameters.}

\item{prior}{A function to provide the log density of prior.}

\item{theta_0}{Initial theta from the prior.}

\item{s_0}{A matrix of the initial summary statistics.}

\item{n_adapt}{Number of initial period.}

\item{n_iter}{Number of iterations.}

\item{sigma_s}{The covariance matrix for kernel function.}

\item{sigma_0}{The covariance matrix for initial period.}

\item{sigma_epi}{Term to control the noise.}

\item{string}{A string for file saving.}
}
\value{
Function value.
}
\description{
Run ABC-MCMC with adaptive Metropolis algorithm.
}
