% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_mcmc_adaptive.R
\name{abc_mcmc_adaptive}
\alias{abc_mcmc_adaptive}
\title{Adaptive ABC-MCMC}
\usage{
abc_mcmc_adaptive(
  obs,
  tol,
  kernel_func,
  p_theta,
  d_theta,
  p_s,
  prior,
  theta_0,
  n_adapt,
  n_iter,
  sigma_0,
  sigma_epi
)
}
\arguments{
\item{obs}{The vector of observed data point.}

\item{tol}{A positive numeric value for the tolerance.}

\item{kernel_func}{A kernel function.}

\item{p_s}{A function to provide the sampled summary statistics by given parameters.}

\item{prior}{A function to provide the log density of prior.}

\item{theta_0}{Initial theta from the prior.}

\item{n_adapt}{Number of initial period.}

\item{n_iter}{Number of iterations.}

\item{sigma_0}{The covariance matrix for initial period.}

\item{sigma_epi}{Term to control the noise.}
}
\value{
Function value.
}
\description{
Run ABC-MCMC with adaptive Metropolis algorithm.
}
