% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARG_igraph_layout.R
\name{ARG_igraph_layout}
\alias{ARG_igraph_layout}
\title{Construct a layout matrix for tree-like ARG}
\usage{
ARG_igraph_layout(ARG)
}
\arguments{
\item{ARG}{An \code{ISM_ARG} or \code{FSM_ARG} object.}
}
\value{
A layout matrix for \code{igraph::plot.igraph}.
}
\description{
Provide the layout coordinates for \code{igraph::plot.igraph}.
}
