% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binary2decimal_cpp}
\alias{binary2decimal_cpp}
\title{Convert binary rows to decimal rows (Rcpp version)}
\usage{
binary2decimal_cpp(x, n, n_last)
}
\arguments{
\item{x}{A boolean vector.}

\item{n}{An integer that devides the length of \code{x}.}

\item{n_last}{An integer as the number of binary elements for the last decimal value.}
}
\value{
A vector of decimal numbers.
}
\description{
For every \code{n} binary numbers, convert them into one decimal number.
}
