% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_to_decimal.R
\name{binary_to_decimal}
\alias{binary_to_decimal}
\title{Convert binary rows to decimal rows}
\usage{
binary_to_decimal(x, n)
}
\arguments{
\item{x}{A boolean vector.}

\item{n}{An integer that devides the length of \code{x}.}
}
\value{
A vector of decimal numbers.
}
\description{
For every \code{n} binary numbers, convert them into one decimal number.
}
