% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_death_sim.R
\name{birth_death_sim}
\alias{birth_death_sim}
\title{Input: n and rho
Simulate a birth death process until it hits the boundary
Output: the time when the process first hit k=1
Hitting time for a birth-death process}
\usage{
birth_death_sim(n, rho)
}
\arguments{
\item{n}{A single integer for initial value.}

\item{rho}{The death rate.}
}
\value{
Hitting time.
}
\description{
Simulate a birth-death process and find its hitting time at boundary equals to 1.
}
\examples{
t <- birth_death_sim(20L, 1)
}
