% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_death_sim2.R
\name{birth_death_sim2}
\alias{birth_death_sim2}
\title{Hitting time for a birth-death process}
\usage{
birth_death_sim2(n, rho)
}
\arguments{
\item{n}{A single integer for initial value.}

\item{rho}{The birth rate.}
}
\value{
Hitting time and number of jumps as a vector.
}
\description{
Simulate a birth-death process and find its hitting time at boundary equals to 1.
}
\examples{
t <- birth_death_sim2(20L, 1)
}
