% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_genealogy.R
\name{clonal_genealogy}
\alias{clonal_genealogy}
\title{Clonal Genealogy simulation}
\usage{
clonal_genealogy(n)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}
}
\value{
A list containing clonal tree.
}
\description{
Simulate coalescent backwards in time to construct a clonal tree.
}
\examples{
tree1 <- clonal_genealogy(10L)
tree2 <- clonal_genealogy(15L)
}
