% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_tree_FSM.R
\name{clonal_tree_FSM}
\alias{clonal_tree_FSM}
\title{Construct clonal tree from ARG}
\usage{
clonal_tree_FSM(ARG)
}
\arguments{
\item{ARG}{An \code{FSM_ARG} object.}
}
\value{
\code{localtree} object; the clonal tree from ARG.
}
\description{
Select clonal lineages to create a clonal tree.
}
\examples{
ARG <- FSM_ARG(20L, 1, 100L)
clonal_tree <- clonal_tree_FSM(ARG)
}
