% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalOrigin_ARG_based.R
\name{ClonalOrigin_ARG_based}
\alias{ClonalOrigin_ARG_based}
\title{An approximated ancestral recombination graph (ARG) using ClonalOrigin algorithm}
\usage{
ClonalOrigin_ARG_based(
  n,
  rho_site,
  L,
  delta,
  node_max = 1000,
  optimise_recomb = FALSE,
  edgemat = TRUE
)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho_site}{The recombination parameter per site.}

\item{L}{An integer for the number of sites.}

\item{delta}{numeric; If bacteria = TRUE, delta is the mean of recombinant segment length.}

\item{node_max}{numeric; Initial maximal node size (default = 1000).}

\item{optimise_recomb}{numeric; If TRUE, skip the recombinations that containing no effective segment.}

\item{edgemat}{numeric; If TRUE, return full edge material matrix.}
}
\value{
A list containing matrices of edges and nodes, and other information about ARG.
}
\description{
Simulate coalescent an recombination events by ClonalOrigin algorithm.
The non-clonal lineages are not allowed to either recombine or coalesce with each other.
}
\examples{
ARG1 <- ClonalOrigin_ARG_based(100L, 0.5, 100L, 5)
ARG2 <- ClonalOrigin_ARG_based(5L, 1, 10L, 1, optimise_recomb=TRUE)
}
