% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalOrigin_nodes.R
\name{ClonalOrigin_nodes}
\alias{ClonalOrigin_nodes}
\title{ClonalOrigin Recombination Nodes}
\usage{
ClonalOrigin_nodes(matrix_data, n)
}
\arguments{
\item{matrix_data}{A recomb_edge matrix.}

\item{n}{The edge index.}
}
\value{
vector of where are the nodes
}
\description{
Find the row index for recombination nodes on a given edge.
}
