% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalOrigin_pair_seq.R
\name{ClonalOrigin_pair_seq}
\alias{ClonalOrigin_pair_seq}
\title{An approximated ancestral recombination graph (ARG) using ClonalOrigin algorithm}
\usage{
ClonalOrigin_pair_seq(tree, rho_site, L, delta, k)
}
\arguments{
\item{tree}{The clonal genealogy.}

\item{rho_site}{The recombination parameter per site.}

\item{L}{An integer for the number of sites.}

\item{delta}{Numeric, delta is the mean of recombinant segment length.}

\item{k}{An integer for the distance between two sites.}
}
\value{
A list containing clonal tree and recombination edges.
}
\description{
Simulate coalescent an recombination events by ClonalOrigin algorithm at two sites.
}
\examples{
ARG1 <- ClonalOrigin_pair_seq(tree, 0.5, 100L, 5, 20L)
ARG2 <- ClonalOrigin_pair_seq(tree, 1, 10L, 1, 5L)
}
