% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalOrigin_pair_seq.simulator_truncated.R
\name{ClonalOrigin_pair_seq.simulator_truncated}
\alias{ClonalOrigin_pair_seq.simulator_truncated}
\title{Simulate approximated ARG using ClonalOrigin with pair method (truncated mutation distribution)}
\usage{
ClonalOrigin_pair_seq.simulator_truncated(
  tree,
  rho_site,
  theta_site,
  L,
  delta,
  N,
  k_vec = c(50L, 200L, 2000L)
)
}
\arguments{
\item{tree}{The clonal genealogy.}

\item{rho_site}{The recombination parameter per site.}

\item{theta_site}{The mutation rate per site.}

\item{L}{An integer for the number of sites.}

\item{delta}{Numeric, delta is the mean of recombinant segment length.}

\item{N}{Number of pairs to simulate.}

\item{k_vec}{A vector for the distance values between two sites. Default is 50, 200, 2000.}
}
\value{
A 7 dimension vector as the summary statistics of simulations.
}
\description{
Simulate pairs of sites in the ClonalOrigin approximation from a given clonal tree.
Provide summary statistics for pairs in three different distance.
}
\examples{
tree <- clonal_genealogy(15L)
summary_stats <- ClonalOrigin_pair_seq.simulator_truncated(tree, 0.5, 0.3, 100L, 5,
                                                           2000, c(10L, 20L, 70L))
}
