% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalOrigin_seq_based.R
\name{ClonalOrigin_seq_based}
\alias{ClonalOrigin_seq_based}
\title{An approximated ancestral recombination graph (ARG) using ClonalOrigin algorithm}
\usage{
ClonalOrigin_seq_based(n, rho, L, delta)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho}{The recombination parameter.}

\item{L}{An integer for the number of sites.}

\item{delta}{numeric; If bacteria = TRUE, delta is the mean of recombinant segment length.}
}
\value{
A list containing clonal tree and recombination edges.
}
\description{
Simulate coalescent an recombination events by ClonalOrigin algorithm.
Recombinations are added to a clonal tree to construct an ARG.
}
\examples{
ARG1 <- ClonalOrigin_seq_based(100L, 5, 100L, 5)
ARG2 <- ClonalOrigin_seq_based(5L, 1, 10L, 1)
}
