% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalOrigin_treetoARG.R
\name{ClonalOrigin_treetoARG}
\alias{ClonalOrigin_treetoARG}
\title{An approximated ancestral recombination graph (ARG) using ClonalOrigin algorithm}
\usage{
ClonalOrigin_treetoARG(n, rho, L, delta, edgemat = TRUE)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho}{The recombination parameter.}

\item{L}{An integer for the number of sites.}

\item{delta}{numeric; If bacteria = TRUE, delta is the mean of recombinant segment length.}

\item{edgemat}{numeric; If TRUE, return full edge material matrix.}
}
\value{
A list containing matrices of edges and nodes, and other information about ARG.
}
\description{
Simulate coalescent an recombination events by ClonalOrigin algorithm.
Recombinations are added to a clonal tree to construct an ARG.
}
\examples{
ARG1 <- ClonalOrigin_treetoARG(100L, 5, 100L, 5)
ARG2 <- ClonalOrigin_treetoARG(5L, 1, 10L, 1, edgemat=FALSE)
}
