% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal2binary.R
\name{decimal2binary}
\alias{decimal2binary}
\title{Convert decimal rows to binary rows}
\usage{
decimal2binary(x, n)
}
\arguments{
\item{x}{A integer vector.}

\item{n}{An integer.}
}
\value{
A boolean vector of binary numbers.
}
\description{
Convert every decimal number to \code{n} binary elements.
}
