% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decimal2binary_cpp}
\alias{decimal2binary_cpp}
\title{Convert decimal rows to binary rows}
\usage{
decimal2binary_cpp(x, n, n_last)
}
\arguments{
\item{x}{A integer vector.}

\item{n}{An integer.}

\item{n_last}{An integer as the number of binary elements for last value in \code{x}.}
}
\value{
A boolean vector of binary numbers.
}
\description{
Convert every decimal number to \code{n} binary elements.
}
