% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_R.R
\name{effective_R}
\alias{effective_R}
\title{Compute the effective recombination rate for \code{simbac_ARG}}
\usage{
effective_R(mat, delta, rho, clonal, include_site = NULL)
}
\arguments{
\item{mat}{A vector of node ancestral material.}

\item{delta}{numeric; The mean of recombinant segment length.}

\item{rho}{The recombination parameter.}

\item{clonal}{logical; whether the lineage is clonal.}

\item{include_site}{A vector of optimised sites.}
}
\value{
Effective recombination rate, cumulative sum of initiating sites probability.
}
\description{
Compute the effective recombination rate,
and provide the probability of recombination initiating sites.
}
