% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSM_ARG.big.R
\name{FSM_ARG.big}
\alias{FSM_ARG.big}
\title{A full ancestral recombination graph (ARG)}
\usage{
FSM_ARG.big(
  n,
  rho,
  L,
  bacteria = FALSE,
  delta = NULL,
  node_max = 1000,
  optimise_recomb = FALSE,
  clonal = FALSE
)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho}{The recombination parameter.}

\item{L}{An integer for the number of sites.}

\item{bacteria}{logical; If TRUE, genes recombine by conversion.}

\item{delta}{numeric; If bacteria = TRUE, delta is the mean of recombinant segment length.}

\item{node_max}{numeric; Initial maximal node size (default = 1000).}

\item{optimise_recomb}{numeric; If TRUE, skip the recombinations that containing no effective segment.}

\item{clonal}{numeric; If TRUE, distinguish clonal lineages.}
}
\value{
A list containing big matrix of nodes, and other information about ARG.
}
\description{
Simulate coalescent an recombination events and construct an ARG.
Assume finite site model. Return a \code{big.matrix} for ancestral material.
}
\examples{
ARG1 <- FSM_ARG.big(20L, 1, 100L)
ARG2 <- FSM_ARG.big(5L, 1, 10L, bacteria = TRUE, delta = 1, optimise_recomb = TRUE, clonal = TRUE)
}
