% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSM_ARG.R
\name{FSM_ARG}
\alias{FSM_ARG}
\title{A full ancestral recombination graph (ARG)}
\usage{
FSM_ARG(
  n,
  rho_site,
  L,
  bacteria = FALSE,
  delta = NULL,
  node_max = 1000,
  optimise_recomb = FALSE,
  clonal = FALSE,
  edgemat = TRUE
)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho_site}{The recombination parameter per site.}

\item{L}{An integer for the number of sites.}

\item{bacteria}{logical; If TRUE, genes recombine by conversion.}

\item{delta}{numeric; If bacteria = TRUE, delta is the mean of recombinant segment length.}

\item{node_max}{numeric; Initial maximal node size (default = 1000).}

\item{optimise_recomb}{numeric; If TRUE, skip the recombinations that containing no effective segment.}

\item{clonal}{numeric; If TRUE, distinguish clonal lineages.}

\item{edgemat}{numeric; If TRUE, return full edge material matrix.}
}
\value{
A list containing matrices of edges and nodes, and other information about ARG.
}
\description{
Simulate coalescent an recombination events and construct an ARG.
Assume finite site model.
}
\examples{
ARG1 <- FSM_ARG(20L, 0.1, 100L)
ARG2 <- FSM_ARG(5L, 1, 10L, bacteria = TRUE, delta = 1, optimise_recomb = TRUE, clonal = TRUE)
}
