% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSM_ARG_mem.R
\name{FSM_ARG_mem}
\alias{FSM_ARG_mem}
\title{Track memory usage of FSM_ARG.decimal}
\usage{
FSM_ARG_mem(
  n,
  rho_site,
  L,
  bacteria = FALSE,
  delta = NULL,
  node_max = 1000,
  optimise_recomb = FALSE,
  clonal = FALSE
)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho_site}{The recombination parameter per site.}

\item{L}{An integer for the number of sites (>= 20).}

\item{bacteria}{logical; If TRUE, genes recombine by conversion.}

\item{delta}{numeric; If bacteria = TRUE, delta is the mean of recombinant segment length.}

\item{node_max}{numeric; Initial maximal node size (default = 1000).}

\item{optimise_recomb}{numeric; If TRUE, skip the recombinations that containing no effective segment.}

\item{clonal}{numeric; If TRUE, distinguish clonal lineages.}
}
\value{
A matrix about memory usage.
}
\description{
Simulate coalescent an recombination events and construct an ARG.
Assume finite site model.
}
\examples{
library(pryr)


ARG1 <- FSM_ARG_mem(20L, 0.1, 100L)
ARG2 <- FSM_ARG_mem(5L, 0.1, 200L, bacteria = TRUE, delta = 10, optimise_recomb = TRUE, clonal = TRUE)
}
