% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSM_mutation.R
\name{FSM_mutation}
\alias{FSM_mutation}
\title{Add mutations to a \code{FSM_ARG} object}
\usage{
FSM_mutation(ARG, theta_site, binary = FALSE)
}
\arguments{
\item{ARG}{A \code{FSM_ARG} object.}

\item{theta_site}{The mutation rate per site.}

\item{binary}{boolean; If TRUE, return the incidence matrix of nodes, instead of sequences.}
}
\value{
Add a mutation matrix and a genotype dataframe to the \code{FSM_ARG} object.
}
\description{
Add mutations uniformly onto the edges of ARG with infinite site assumption.
}
\examples{
ARG <- FSM_ARG(20L, 1, 100L)
ARG_mutation <- FSM_mutation(ARG, 2)
}
