% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSM_mutation_fast.R
\name{FSM_mutation_fast}
\alias{FSM_mutation_fast}
\title{Add mutations to a \code{FSM_ARG} object}
\usage{
FSM_mutation_fast(ARG, theta_site)
}
\arguments{
\item{ARG}{A \code{FSM_ARG} object.}

\item{theta_site}{The mutation rate per site.}
}
\value{
Add a mutation matrix and a genotype dataframe to the \code{FSM_ARG} object.
}
\description{
Add mutations uniformly onto the edges of ARG with infinite site assumption.
Return only the incidence matrix of nodes.
}
\details{
No mutation matrix and only keep \code{mutate_edge} and \code{mutate_site}.
No \code{node} dataframe, simulate \code{node_site} through the tree.
}
\examples{
ARG <- FSM_ARG(20L, 1, 100L)
ARG_mutation <- FSM_mutation_fast(ARG, 2)
}
