% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G3_test.R
\name{G3_test}
\alias{G3_test}
\title{Three-gamete test}
\usage{
G3_test(mat)
}
\arguments{
\item{mat}{An incidence matrix with two columns.}
}
\value{
proportion of combinations.
}
\description{
Compute the proportion of combinations that have patterns 01, 10, 11.
}
\examples{
r_square <- G3_test(mat)
}
