% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_kernel.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Gaussian kernel function}
\usage{
gaussian_kernel(y, z, tol = 1, sigma, log.kernel = TRUE)
}
\arguments{
\item{y}{One vector.}

\item{z}{One vector.}

\item{tol}{The tolerance epsilon.}

\item{sigma}{The covariance matrix.}

\item{log.kernel}{If TRUE, return value in log scale.}
}
\value{
Function value.
}
\description{
Give the function value for a Gaussian kernel.
}
