% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoplasy_index.R
\name{homoplasy_index}
\alias{homoplasy_index}
\title{Homoplasy index}
\usage{
homoplasy_index(ARG)
}
\arguments{
\item{ARG}{A \code{FSM_ARG} object after mutation.}
}
\value{
A numerical value, homoplasy index.
}
\description{
Compute homoplasy index from ARG and sequence data.
}
\examples{
tree <- clonal_genealogy(15L)
ARG <- ClonalOrigin_pair_seq_fast(tree, 0.5, 100L, 5, c(10L, 20L, 70L))
ARG_mutated <- FSM_mutation_fast(ARG, 0.3)
hi <- homoplasy_index(ARG_mutated)
}
