% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISM_mutation.R
\name{ISM_mutation}
\alias{ISM_mutation}
\title{Add mutations to an \code{ISM_ARG} object}
\usage{
ISM_mutation(ARG, theta)
}
\arguments{
\item{ARG}{An \code{ISM_ARG} object.}

\item{theta}{The mutation rate.}
}
\value{
Add a mutation dataframe and a new column of genotype in node dataframe to the \code{ISM_ARG} object.
}
\description{
Add mutations uniformly onto the edges of ARG with infinite site assumption.
}
\examples{
ARG <- ISM_ARG(5L, 1)
ARG_mutation <- ISM_mutation(ARG, 2)
}
