% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_r.R
\name{LD_r}
\alias{LD_r}
\title{The square of correlation coefficient for LD}
\usage{
LD_r(mat)
}
\arguments{
\item{mat}{An incidence matrix with two columns.}
}
\value{
r square.
}
\description{
Compute the square of Pearson’s correlation coefficient for linkage disequilibrium.
}
\examples{
r_square <- LD(mat)
}
