% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_height_ClonalOrigin.R
\name{local_height_ClonalOrigin}
\alias{local_height_ClonalOrigin}
\title{Local tree height for \code{ClonalOrigin} object}
\usage{
local_height_ClonalOrigin(ARG, location)
}
\arguments{
\item{ARG}{A \code{ClonalOrigin} object.}

\item{location}{The site location for local tree.}
}
\value{
Height of the local tree
}
\description{
Find the height of a locat site at give site for a \code{ClonalOrigin} object.
}
