% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_tree.R
\name{local_tree}
\alias{local_tree}
\title{Construct local tree from ARG}
\usage{
local_tree(ARG, location)
}
\arguments{
\item{ARG}{An \code{ISM_ARG} or \code{FSM_ARG} object.}

\item{location}{The site location for local tree.}
}
\value{
\code{localtree} object; Local tree at a chosen site.
}
\description{
Call function \code{local_tree_ISM} or \code{local_tree_FSM}.
Select edges by a chosen site to create the local tree.
}
\examples{
ARG <- FSM_ARG(20L, 1, 100L)
local_tree <- local_tree(ARG_mutation, 2L)
}
