% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localtree_to_phylo.R
\name{localtree_to_phylo}
\alias{localtree_to_phylo}
\title{Input: tree (localtree), label
Convert the local tree to a phylo object that can be plotted by ape::plot.phylo
Output: a phylo object
Convert a \code{localtree} object to a \code{phylo} object}
\usage{
localtree_to_phylo(tree, label = FALSE)
}
\arguments{
\item{tree}{A \code{localtree} object.}

\item{label}{logical; If TRUE, add string labels on leaf lineages.}
}
\value{
\code{phylo} object in \code{ape} package.
}
\description{
Change a local tree to a \code{phylo} object in \code{ape} package.
}
\examples{
ARG <- FSM_ARG(20L, 1, 100L)
local_tree <- local_tree(ARG_mutation, 2L)
phylo_tree1 <- localtree_to_phylo(local_tree, label=TRUE)
}
