% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ARG.R
\name{plot.ARG}
\alias{plot.ARG}
\title{Plot a tree-like ARG}
\usage{
\method{plot}{ARG}(ARG)
}
\arguments{
\item{ARG}{An \code{ISM_ARG} or \code{FSM_ARG} object.}
}
\value{
A plot from \code{igraph::plot.igraph}.
}
\description{
\code{plot.ARG} is able to plot a tree-like network for an ARG.
}
\examples{
ARG <- FSM_ARG(20L, 1, 100L)
plot.ARG(ARG)
}
