% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSM_mutation.R
\name{sim_FSM_mutation}
\alias{sim_FSM_mutation}
\title{Add mutations to an \code{FSM_ARG} object}
\usage{
sim_FSM_mutation(ARG, theta)
}
\arguments{
\item{ARG}{An \code{FSM_ARG} object.}

\item{theta}{The mutation rate.}
}
\value{
Add a mutation matrix and a genotype dataframe to the \code{FSM_ARG} object.
}
\description{
Add mutations uniformly onto the edges of ARG with infinite site assumption.
}
\examples{
ARG <- FSM_ARG(20L, 1, 100L)
ARG_mutation <- ISM_mutation(ARG, 2)
}
