% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbac_ARG.decimal.R
\name{simbac_ARG.decimal}
\alias{simbac_ARG.decimal}
\title{A full ancestral recombination graph (ARG) using simbac algorithm}
\usage{
simbac_ARG.decimal(
  n,
  rho_site,
  L,
  delta,
  node_max = 1000,
  output_eff_R = FALSE,
  edgemat = TRUE
)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho_site}{The recombination parameter per site.}

\item{L}{An integer for the number of sites.}

\item{delta}{Numeric, delta is the mean of recombinant segment length.}

\item{node_max}{Numeric; Initial maximal node size (default = 1000).}

\item{output_eff_R}{Logical; If TRUE, return effective recombination rates.}

\item{edgemat}{Numeric; If TRUE, return full edge material matrix.}
}
\value{
A list containing matrices of edges and nodes, and other information about ARG.
}
\description{
Simulate coalescent an recombination events by considering effective recombination rate.
Assume finite site model.
}
\examples{
ARG1 <- simbac_ARG.decimal(100L, 0.1, 100L, 5)
ARG2 <- simbac_ARG.decimal(5L, 0.1, 10L, 1, output_eff_R = TRUE)
}
