% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbac_ARG.R
\name{simbac_ARG}
\alias{simbac_ARG}
\title{Create a full ancestral recombination graph (ARG) using simbac algorithm}
\usage{
simbac_ARG(
  n,
  rho,
  L,
  delta,
  node_max = 1000,
  output_eff_R = FALSE,
  optimise_site = FALSE,
  edgemat = TRUE
)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho}{The recombination parameter.}

\item{L}{An integer for the number of sites.}

\item{delta}{numeric; If bacteria = TRUE, delta is the mean of recombinant segment length.}

\item{node_max}{numeric; Initial maximal node size (default = 1000).}

\item{output_eff_R}{logical; If TRUE, return effective recombination rates.}

\item{optimise_site}{logical; If TRUE, optimise coalesced sites.}

\item{edgemat}{numeric; If TRUE, return full edge material matrix.}
}
\value{
A list containing matrices of edges and nodes, and other information about ARG.
}
\description{
Simulate coalescent an recombination events by considering effective recombination rate.
Assume finite site model.
}
\examples{
ARG1 <- simbac_ARG(100L, 5, 100L, 5)
ARG2 <- simbac_ARG(5L, 1, 10L, 1, output_eff_R = TRUE, optimise_site = TRUE)
}
