% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_am.R
\name{sl_am}
\alias{sl_am}
\title{Synthetic Likelihood adaptive M-H}
\usage{
sl_am(
  M,
  iter,
  burn_in,
  obs,
  init_theta,
  prior_func,
  sample_func,
  q_sigma,
  epsilon,
  acc_rate = FALSE
)
}
\arguments{
\item{M}{Number of new data points drawn in each iteration.}

\item{iter}{Number of iterations.}

\item{burn_in}{Length of the burn-in period.}

\item{obs}{A vector of the observed statistics.}

\item{init_theta}{A vector of the initial parameter sampled from the prior.}

\item{prior_func}{A density function of prior (log density).}

\item{sample_func}{A function which takes theta and M and return sample mean and variance.}

\item{q_sigma}{A scale matrix for the Gaussian proposal.}

\item{epsilon}{Term to control the noise.}

\item{acc_rate}{Default acc_rate = FALSE, if TRUE, print acceptance rate and return it.}
}
\value{
A sequence of parameters from the BSL posterior.
}
\description{
Apply adaptive Metropolis-Hastings algorithm for BSL posterior as the target distribution.
}
