% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_mcmc.R
\name{sl_mcmc}
\alias{sl_mcmc}
\title{Synthetic Likelihood MCMC}
\usage{
sl_mcmc(
  M,
  iter,
  obs,
  init_theta,
  prior_func,
  sample_func,
  proposal,
  acc_rate = FALSE
)
}
\arguments{
\item{M}{Number of new data points drawn in each iteration.}

\item{iter}{Number of iterations.}

\item{obs}{A vector of the observed statistics.}

\item{init_theta}{A vector of the initial parameter sampled from the prior.}

\item{prior_func}{A density function of prior (log density).}

\item{sample_func}{A function which takes theta and M and return sample mean and variance.}

\item{proposal}{A function which takes theta_old and return theta_new.}

\item{acc_rate}{Default acc_rate = FALSE, if TRUE, print acceptance rate and return it.}
}
\value{
A sequence of parameters from the BSL posterior.
}
\description{
Apply Metropolis-Hastings MCMC (MH-MCMC) algorithm for BSL posterior as the target distribution.
}
