% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_height.R
\name{tree_height}
\alias{tree_height}
\title{Local tree height}
\usage{
tree_height(tree)
}
\arguments{
\item{tree}{An \code{localtree} object.}
}
\value{
Tree height.
}
\description{
Compute the length from leaf node to root node in the local tree.
}
