% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_mcmc.R, R/abc_mcmc_adaptive.R
\name{abc_mcmc}
\alias{abc_mcmc}
\title{ABC-MCMC}
\usage{
abc_mcmc(
  obs,
  tol,
  kernel_func,
  p_theta,
  d_theta,
  p_s,
  prior,
  theta_0,
  n_adapt,
  n_iter,
  sigma_0
)

abc_mcmc(
  obs,
  tol,
  kernel_func,
  p_theta,
  d_theta,
  p_s,
  prior,
  theta_0,
  n_adapt,
  n_iter,
  sigma_0
)
}
\arguments{
\item{obs}{The vector of observed data point.}

\item{tol}{A positive numeric value for the tolerance.}

\item{kernel_func}{A kernel function.}

\item{p_theta}{A function to provide sampled theta from proposal.}

\item{d_theta}{A function to provide the log density of proposal.}

\item{p_s}{A function to provide the sampled summary statistics by given parameters.}

\item{prior}{A function to provide the log density of prior.}

\item{theta_0}{Initial theta from the prior.}

\item{n_adapt}{Number of initial period.}

\item{n_iter}{Number of iterations.}

\item{sigma}{The covariance matrix.}
}
\value{
Function value.

Function value.
}
\description{
Run ABC-MCMC by given proposal and prior.

Run ABC-MCMC with adaptive Metropolis algorithm.
}
