% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalOrigin_tree_seq.R
\name{ClonalOrigin_tree_seq}
\alias{ClonalOrigin_tree_seq}
\title{An approximated ancestral recombination graph (ARG) using ClonalOrigin algorithm}
\usage{
ClonalOrigin_tree_seq(n, rho_site, L, delta)
}
\arguments{
\item{n}{An integer for the number of leaf lineages.}

\item{rho_site}{The recombination parameter per site.}

\item{L}{An integer for the number of sites.}

\item{delta}{Numeric, delta is the mean of recombinant segment length.}
}
\value{
A list containing clonal tree and recombination edges.
}
\description{
Simulate coalescent an recombination events by ClonalOrigin algorithm.
Recombinations are added to a clonal tree to construct an ARG.
}
\examples{
ARG1 <- ClonalOrigin_tree_seq(100L, 0.5, 100L, 5)
ARG2 <- ClonalOrigin_tree_seq(5L, 1, 10L, 1)
}
